Momento System + New Sprites
Added a detailed trinket/momento system.

Added a lot of foliage sprites, will add colliders to all of them slowly over time.

+++ Assets/Scripts/HUDController.cs
@@ -89,7 +89,7 @@ public class HUDController : MonoBehaviour
     /// <summary>
     /// Updates the player's health on the HUD.
     /// </summary>
-    public void UpdateHealthUI(int currentHealth, int maxHealth)
+    public void UpdateHealthUI(float currentHealth, float maxHealth)
     {
         Debug.Log("Healthed");
         if (healthText != null)
