+++ Assets/Scripts/Scriptable Objects/WeaponData.cs
@@ -22,9 +22,9 @@ public class WeaponData : ScriptableObject
     public float fireRate = 0.1f; // Time between shots (Automatic fire only)
     public float spread = 0.1f; // Bullet inaccuracy (0 = no spread)
     public int bulletsPerShot = 1; // Number of bullets fired per shot (e.g., for shotguns)
-    public int damage = 10; // Damage per bullet
+    public float damage = 10; // Damage per bullet
     public AmmoType ammoType; // Choose the ammo type.... be careful, you can make the shotgun shoot rockets if you wanted it to.
-    public int apValue; // armor piercing value, i.e how good is it at piercing armor
+    public float apValue; // armor piercing value, i.e how good is it at piercing armor
 
     [TextArea]
     public string description; // Optional description for the weapon
